backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.35.16
botocore==1.35.16
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distro==1.9.0
ecdsa==0.19.0
effdet==0.4.1
emoji==2.10.1
et-xmlfile==1.1.0
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.13
langchain-cli==0.0.21
langchain-community==0.0.31
langchain-core==0.1.40
langchain-experimental==0.0.55
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langdetect==1.0.9
langserve==0.0.51
langsmith==0.1.38
layoutparser==0.3.4
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.16.2
opencv-python==4.8.0.76
openpyxl==3.1.2
orjson==3.10.0
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pgvector==0.2.5
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
protobuf==4.23.4
psycopg==3.1.18
psycopg2==2.9.9
pyasn1==0.6.1
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.3
Pygments==2.17.2
pypandoc==1.12
pyparsing==3.0.9
pypdf==4.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.3.2
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==1.8.2
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
timm==0.9.12
tokenizers==0.15.2
tomlkit==0.12.4
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
